refactor: enhance integration test setup by spliting client and kit setup in different files
